rename thermal -> heat_transfer
